# cmd/controller/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Qryptic Controller Service to manage the users, groups, gateways and clients.",
        "title": "Qryptic Controller API",
        "contact": {
            "name": "Leetsecure Support",
            "url": "leetsecure.com",
            "email": "hello@leetsecure.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/client/{id}": {
            "delete": {
                "description": "DeleteVpnClient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "DeleteVpnClient",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "client id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/get-gateway-config": {
            "get": {
                "description": "GetVpnGatewayWGConfigByGW",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateway"
                ],
                "summary": "GetVpnGatewayWGConfigByGW",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGServerConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/list": {
            "get": {
                "description": "GetVpnGatewaysAccessibleByUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "GetVpnGatewaysAccessibleByUser",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.VpnGatewayUserResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/gateway/{id}/client": {
            "get": {
                "description": "GetVpnClientConfig",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "GetVpnClientConfig",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "gateway id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGClientConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.VpnGatewayUserResponse": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "serverPublicKey": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGClientConfig": {
            "type": "object",
            "properties": {
                "clientInterfaceConfig": {
                    "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGClientInterfaceConfig"
                },
                "clientPeerConfig": {
                    "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGClientPeerConfig"
                },
                "clientUuid": {
                    "type": "string"
                },
                "expiryTime": {
                    "type": "string"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGClientInterfaceConfig": {
            "type": "object",
            "properties": {
                "dnsServer": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGClientPeerConfig": {
            "type": "object",
            "properties": {
                "allowedIPs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "persistantAlive": {
                    "type": "integer"
                },
                "presharedKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                },
                "vpnGatewayDomain": {
                    "type": "string"
                },
                "vpnGatewayIP": {
                    "type": "string"
                },
                "vpnGatewayPort": {
                    "type": "integer"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGServerConfig": {
            "type": "object",
            "properties": {
                "wgServerInterfaceConfig": {
                    "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGServerInterfaceConfig"
                },
                "wgServerPeerConfigs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGServerPeerConfig"
                    }
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGServerInterfaceConfig": {
            "type": "object",
            "properties": {
                "dnsServer": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "listenPort": {
                    "type": "integer"
                },
                "postDown": {
                    "type": "string"
                },
                "postUp": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                },
                "vpnGatewayUuid": {
                    "type": "string"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.WGServerPeerConfig": {
            "type": "object",
            "properties": {
                "clientAllowedIPs": {
                    "type": "string"
                },
                "clientPublicKey": {
                    "type": "string"
                },
                "presharedKey": {
                    "type": "string"
                }
            }
        },
...
    }
}